+++ src/misc.c
@@ -2146,7 +2146,7 @@ void *rig_get_function_ptr(rig_model_t rig_model,
  * \param RIG* and rig_caps_int_e
  * \return the corresponding long value -- -RIG_EINVAL is the only error possible
  */
-long rig_get_caps_int(rig_model_t rig_model, enum rig_caps_int_e rig_caps)
+long long rig_get_caps_int(rig_model_t rig_model, enum rig_caps_int_e rig_caps)
 {
     const struct rig_caps *caps = rig_get_caps(rig_model);
